+++ starters/validation/src/config/index.ts
@@ -1,0 +1,1 @@
+export {ValidationsConfiguration} from "./ValidationsConfiguration";