+++ starters/validation/src/decorator/EnableValidations.ts
@@ -1,0 +1,7 @@
+import {ValidationsConfiguration} from "../config";
+
+export const EnableValidations = (): ClassDecorator => {
+    return () => {
+        new ValidationsConfiguration();
+    };
+};