+++ starters/validation/src/decorator/index.ts
@@ -1,0 +1,1 @@
+export {EnableValidations} from "./EnableValidations";