+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/config
 
+## 1.1.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.3.1
+
 ## 1.1.3
 
 ### Patch Changes