+++ starters/openapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/starter-openapi
 
+## 1.6.2
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/engine@1.1.4
+    -   @nodeboot/context@1.3.1
+
 ## 1.6.1
 
 ### Patch Changes