+++ starters/scheduler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/starter-scheduler
 
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.3.1
+
 ## 1.0.3
 
 ### Patch Changes